requirements.txt
pyperclip
pillow
pyscreenshot
requests
keyboard